+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed
+- `GET /login` handler checks if `module:@authnomicon/credentials.PasswordStore`
+interface is implemented prior to redirecting to `/login/password`.  If not, the
+original render error is propagated.
 - Moved `login/http/prompt`, and associated files, to `prompt`.
 - Moved `login/http/service`, and associated files, to `service`.